requirements.txt
appnope==0.1.4
asttokens==3.0.1
autograd==1.8.0
certifi==2025.11.12
charset-normalizer==3.4.4
comm==0.2.3
contourpy==1.3.2
cramjam==2.11.0
cycler==0.12.1
debugpy==1.8.17
decorator==5.2.1
dill==0.4.0
ebaysdk==2.2.0
et_xmlfile==2.0.0
exceptiongroup==1.3.1
executing==2.2.1
fastparquet==2025.12.0
filelock==3.20.1
fonttools==4.61.0
fsspec==2025.12.0
hf-xet==1.2.0
huggingface-hub==0.36.0
idna==3.11
ipykernel==7.1.0
ipython==8.37.0
jedi==0.19.2
Jinja2==3.1.6
joblib==1.5.2
jupyter_client==8.6.3
jupyter_core==5.9.1
kiwisolver==1.4.9
kmedoids==0.5.4
Lifetimes==0.11.3
lxml==6.0.2
MarkupSafe==3.0.3
matplotlib==3.10.7
matplotlib-inline==0.2.1
mpmath==1.3.0
nest-asyncio==1.6.0
networkx==3.4.2
numpy==2.2.6
openpyxl==3.1.5
packaging==25.0
pandas==2.3.3
parso==0.8.5
patsy==1.0.2
pexpect==4.9.0
pillow==12.0.0
platformdirs==4.5.1
prompt_toolkit==3.0.52
psutil==7.1.3
ptyprocess==0.7.0
pure_eval==0.2.3
pyarrow==22.0.0
Pygments==2.19.2
pyparsing==3.2.5
python-dateutil==2.9.0.post0
pytz==2025.2
PyYAML==6.0.3
pyzmq==27.1.0
regex==2025.11.3
requests==2.32.5
safetensors==0.7.0
scikit-learn==1.7.2
scikit-learn-extra==0.3.0
scipy==1.15.3
seaborn==0.13.2
sentence-transformers==5.2.0
six==1.17.0
stack-data==0.6.3
statsmodels==0.14.6
sympy==1.14.0
threadpoolctl==3.6.0
tokenizers==0.22.1
torch==2.9.1
tornado==6.5.2
tqdm==4.67.1
traitlets==5.14.3
transformers==4.57.3
typing_extensions==4.15.0
tzdata==2025.2
urllib3==2.6.1
wcwidth==0.2.14